Adding startup scripts for multiple simulators with application settings for iOS, making multiple files for ios, android.
