+++ lib/parallel_calabash.rb
@@ -1,8 +1,10 @@
 require 'tempfile'
 require 'parallel'
 require 'parallel_calabash/version'
-require 'parallel_calabash/adb_helper'
-require 'parallel_calabash/runner'
+require 'parallel_calabash/android_helper'
+require 'parallel_calabash/ios_helper'
+require 'parallel_calabash/android_runner'
+require 'parallel_calabash/ios_runner'
 require 'parallel_calabash/feature_grouper'
 require 'parallel_calabash/result_formatter'
 require 'rbconfig'
