+++ lib/parallel_calabash/android_helper.rb
@@ -1,0 +1,38 @@
+require 'parallel_calabash/device_helper'
+
+module ParallelCalabash
+  class AdbHelper
+    include ParallelCalabash::DevicesHelper
+
+    def initialize(filter = [])
+      @filter = filter
+    end
+
+    def connected_devices_with_model_info
+      begin
+        list =
+            `adb devices -l`.split("\n").collect do |line|
+              device = device_id_and_model(line)
+              filter_device(device)
+            end
+        list.compact
+      rescue
+        []
+      end
+    end
+
+    def device_id_and_model line
+      if line.match(/device(?!s)/)
+        [line.split(" ").first, line.scan(/model:(.*) device/).flatten.first]
+      end
+    end
+
+    def filter_device device
+      if @filter && !@filter.empty? && device
+        device unless @filter.collect { |f| device[0].match(f) || device[1].match(f) }.compact.empty?
+      else
+        device
+      end
+    end
+  end
+end
