+++ lib/parallel_calabash/device_helper.rb
@@ -1,0 +1,13 @@
+require 'fileutils'
+
+module ParallelCalabash
+  module DevicesHelper
+    def device_for_process process_num
+      connected_devices_with_model_info[process_num]
+    end
+
+    def number_of_connected_devices
+      connected_devices_with_model_info.size
+    end
+  end
+end
