+++ lib/parallel_calabash/version.rb
@@ -1,3 +1,3 @@
 module ParallelCalabash
-  VERSION = "0.2.2"
+  VERSION = "0.2.3-SNAPSHOT"
 end
